+++ src/Sulu/Bundle/MediaBundle/Api/Collection.php
@@ -10,16 +10,17 @@
 
 namespace Sulu\Bundle\MediaBundle\Api;
 
-use Sulu\Bundle\MediaBundle\Entity\CollectionInterface;
-use JMS\Serializer\Annotation\VirtualProperty;
-use JMS\Serializer\Annotation\SerializedName;
-use Sulu\Bundle\MediaBundle\Entity\CollectionMeta;
-use Sulu\Bundle\MediaBundle\Entity\CollectionType;
-use JMS\Serializer\Annotation\ExclusionPolicy;
-use Sulu\Component\Rest\ApiWrapper;
-use Sulu\Component\Security\Authentication\UserInterface;
+use Hateoas\Configuration\Annotation\Embedded;
 use Hateoas\Configuration\Annotation\Relation;
 use Hateoas\Configuration\Annotation\Route;
+use JMS\Serializer\Annotation\ExclusionPolicy;
+use JMS\Serializer\Annotation\SerializedName;
+use JMS\Serializer\Annotation\VirtualProperty;
+use Sulu\Bundle\MediaBundle\Entity\CollectionInterface;
+use Sulu\Bundle\MediaBundle\Entity\CollectionMeta;
+use Sulu\Bundle\MediaBundle\Entity\CollectionType;
+use Sulu\Component\Rest\ApiWrapper;
+use Sulu\Component\Security\Authentication\UserInterface;
 
 /**
  * Class Collection
@@ -40,18 +41,72 @@ use Hateoas\Configuration\Annotation\Route;
  *          parameters = { "collection" = "expr(object.getId())", "limit" = 9999, "types" = "{types}" }
  *      )
  * )
+ * @Relation(
+ *      "self",
+ *      href = @Route(
+ *          "get_collection",
+ *          parameters = { "id" = "expr(object.getId())" }
+ *      )
+ * )
+ * @Relation(
+ *      "admin",
+ *      href ="expr('media/collections/edit:'~object.getId()~'/files')"
+ * )
+ * @Relation(
+ *      "children",
+ *      href = @Route(
+ *          "get_collection",
+ *          parameters = { "id" = "expr(object.getId())", "depth" = 1 }
+ *      )
+ * )
+ * @Relation(
+ *     name = "collections",
+ *     embedded = @Embedded(
+ *         "expr(object.getChildren())",
+ *         xmlElementName = "collections"
+ *     )
+ * )
+ * @Relation(
+ *     name = "parent",
+ *     embedded = @Embedded(
+ *         "expr(object.getParent())",
+ *         xmlElementName = "parent"
+ *     )
+ * )
+ * @Relation(
+ *     name = "breadcrumb",
+ *     embedded = @Embedded(
+ *         "expr(object.getBreadcrumb())",
+ *         xmlElementName = "breadcrumb"
+ *     )
+ * )
  */
 class Collection extends ApiWrapper
 {
     /**
      * @var array
      */
-    protected $previews = array();
+    protected $preview = array();
 
     /**
      * @var array
      */
     protected $properties = array();
+
+    /**
+     * @var Collection[]
+     */
+    protected $children = array();
+
+    /**
+     * @var Collection
+     */
+    protected $parent;
+
+    /**
+     * @var array
+     */
+    protected $breadcrumb;
 
     public function __construct(CollectionInterface $collection, $locale)
     {
@@ -60,19 +115,46 @@ class Collection extends ApiWrapper
     }
 
     /**
-     * @VirtualProperty
-     * @SerializedName("children")
-     * @return array
+     * Set children of resource
+     * @param Collection[] $children
+     */
+    public function setChildren($children)
+    {
+        $childrenEntities = array();
+
+        foreach ($children as $child) {
+            $childrenEntities[] = $child->getEntity();
+        }
+
+        // FIXME remove cache for children and generate then on the fly
+        //       reason: preview images cannot be generated without a service
+        $this->children = $children;
+    }
+
+    /**
+     * Returns children of resource
+     * @return Collection[]
      */
     public function getChildren()
     {
-        $childIds = array();
-        /** @var Entity $child */
-        foreach ($this->entity->getChildren() as $child) {
-            $childIds[] = $child->getId();
+        // FIXME remove cache for children and generate then on the fly
+        //       reason: preview images cannot be generated without a service
+        return $this->children;
+    }
+
+    /**
+     * Indicates if sub collections exists
+     * @VirtualProperty
+     * @SerializedName("hasSub")
+     * @return boolean
+     */
+    public function getHasSub()
+    {
+        if (($children = $this->getEntity()->getChildren()) !== null) {
+            return $children->count() > 0;
         }
 
-        return $childIds;
+        return false;
     }
 
     /**
@@ -112,44 +194,48 @@ class Collection extends ApiWrapper
     }
 
     /**
-     * @VirtualProperty
-     * @SerializedName("mediaNumber")
-     * @return int
-     */
-    public function getMediaNumber()
-    {
-        $mediaCount = 0;
-
-        /** @var Entity $child */
-        foreach ($this->entity->getChildren() as $child) {
-            $mediaCount += count($child->getMedia());
-        }
-
-        // set media count
-        $mediaCount += count($this->entity->getMedia());
-
-        return $mediaCount;
-    }
-
-    /**
-     * @param Entity $parent
+     * @param Collection $parent
      * @return $this
      */
     public function setParent($parent)
     {
-        $this->entity->setParent($parent);
+        if ($parent !== null) {
+            $this->entity->setParent($parent->getEntity());
+        } else {
+            $this->entity->setParent(null);
+        }
+
+        // FIXME remove cache for parent and generate it on the fly
+        //       reason: preview images cannot be generated without a service
+        $this->parent = $parent;
 
         return $this;
     }
 
     /**
-     * @VirtualProperty
-     * @SerializedName("parent")
-     * @return int|null
+     * @return Collection
      */
     public function getParent()
     {
-        return ($parent = $this->entity->getParent()) ? $parent->getId() : null;
+        // FIXME remove cache for parent and generate it on the fly
+        //       reason: preview images cannot be generated without a service
+        return $this->parent;
+    }
+
+    /**
+     * @return array
+     */
+    public function getBreadcrumb()
+    {
+        return $this->breadcrumb;
+    }
+
+    /**
+     * @param array $breadcrumb
+     */
+    public function setBreadcrumb($breadcrumb)
+    {
+        $this->breadcrumb = $breadcrumb;
     }
 
     /**
@@ -198,29 +284,22 @@ class Collection extends ApiWrapper
     }
 
     /**
+     * @VirtualProperty
+     * @SerializedName("preview")
      * @return array
      */
-    public function getPreviews()
+    public function getPreview()
     {
-        return $this->previews;
+        return $this->preview;
     }
 
     /**
-     * @VirtualProperty
-     * @SerializedName("thumbnails")
-     */
-    public function getThumbnails() // FIXME change to getPreviews when SerializedName working
-    {
-        return $this->previews;
-    }
-
-    /**
-     * @param array $previews
+     * @param array $preview
      * @return $this
      */
-    public function setPreviews($previews)
+    public function setPreview($preview)
     {
-        $this->previews = $previews;
+        $this->preview = $preview;
 
         return $this;
     }
@@ -364,14 +443,16 @@ class Collection extends ApiWrapper
         $metaCollection = $this->entity->getMeta();
 
         // get meta only with this locale
-        $metaCollectionFiltered = $metaCollection->filter(function ($meta) use ($locale) {
-            /** @var CollectionMeta $meta */
-            if ($meta->getLocale() == $locale) {
-                return true;
-            }
+        $metaCollectionFiltered = $metaCollection->filter(
+            function ($meta) use ($locale) {
+                /** @var CollectionMeta $meta */
+                if ($meta->getLocale() == $locale) {
+                    return true;
+                }
 
-            return false;
-        });
+                return false;
+            }
+        );
 
         // check if meta was found
         if ($metaCollectionFiltered->isEmpty()) {