+++ src/Sulu/Bundle/MediaBundle/Controller/TemplateController.php
@@ -63,14 +63,4 @@ class TemplateController extends Controller
     {
         return $this->render('SuluMediaBundle:Template:media-multiple-edit.html.twig');
     }
-
-    /**
-     * renders template for the collection list
-     * @return Response
-     */
-    public function collectionListAction()
-    {
-        return $this->render('SuluMediaBundle:Template:collection-list.html.twig');
-    }
-
 }