+++ src/Sulu/Bundle/MediaBundle/DependencyInjection/SuluMediaExtension.php
@@ -43,8 +43,7 @@ class SuluMediaExtension extends Extension
         $container->setParameter('sulu_media.collection.type.default', array(
             'id' => 1
         ));
-        $container->setParameter('sulu_media.collection.previews.limit', 3);
-        $container->setParameter('sulu_media.collection.previews.format', '150x100');
+        $container->setParameter('sulu_media.collection.previews.format', '50x50');
         $container->setParameter('sulu_media.media.max_file_size', '16MB');
         $container->setParameter('sulu_media.media.blocked_file_types', array('file/exe'));
         $container->setParameter('sulu_media.media.storage.local.path', '%kernel.root_dir%/../uploads/media');