+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/collections/components/content/main.js
@@ -7,57 +7,134 @@
  * with this source code in the file LICENSE.
  */
 
-define(function () {
+define(function() {
 
     'use strict';
 
     var namespace = 'sulu.media.collections-edit.',
 
-    /**
-     * sets the locale
-     * @event sulu.media.collections.collection-list
-     * @param {String} the new locale
-     */
-    SET_LOCALE = function () {
-        return createEventName.call(this, 'set-locale');
-    },
-
-    /**
-     * sets the locale
-     * @event sulu.media.collections.collection-list
-     * @param {Function} the callback to pass the locale to
-     */
-    GET_LOCALE = function () {
-        return createEventName.call(this, 'get-locale');
-    },
-
-    /** returns normalized event names */
-    createEventName = function (postFix) {
-        return namespace + postFix;
-    };
-
-    return {
-        header: {
-            tabs: {
-                url: '/admin/media/navigation/collection'
-            }
+        /**
+         * sets the locale
+         * @event sulu.media.collections.collection-list
+         * @param {String} the new locale
+         */
+        SET_LOCALE = function() {
+            return createEventName.call(this, 'set-locale');
         },
 
-        initialize: function () {
+        /**
+         * sets the locale
+         * @event sulu.media.collections.collection-list
+         * @param {Function} the callback to pass the locale to
+         */
+        GET_LOCALE = function() {
+            return createEventName.call(this, 'get-locale');
+        },
+
+        /** returns normalized event names */
+        createEventName = function(postFix) {
+            return namespace + postFix;
+        };
+
+    return {
+        header: function() {
+            var def = this.sandbox.data.deferred();
+            this.sandbox.util.load('/admin/api/localizations')
+                .then(function(data) {
+                    var localizations = data.map(function(localization) {
+                        return {
+                            id: localization.localization,
+                            title: localization.localization
+                        };
+                    });
+
+                    def.resolve({
+                        tabs: {
+                            url: '/admin/media/navigation/collection'
+                        },
+                        toolbar: {
+                            template: [
+                                {
+                                    id: 'settings',
+                                    icon: 'gear',
+                                    position: 30,
+                                    items: [
+                                        {
+                                            id: 'collection-move',
+                                            title: this.sandbox.translate('sulu.collection.move'),
+                                            callback: this.startMoveCollectionOverlay.bind(this)
+                                        },
+                                        {
+                                            id: 'delete',
+                                            title: this.sandbox.translate('sulu.collections.delete-collection'),
+                                            callback: this.deleteCollection.bind(this)
+                                        }
+                                    ]
+                                }
+                            ],
+                            parentTemplate: 'save',
+                            languageChanger: {
+                                data: localizations,
+                                preSelected: this.options.locale
+                            }
+                        },
+                        noBack: true
+                    })
+                }.bind(this));
+
+            return def;
+        },
+
+        /**
+         * Deletes the current collection
+         */
+        deleteCollection: function() {
+            this.sandbox.emit('sulu.media.collections.delete-collection', this.options.id);
+        },
+
+        /**
+         * starts overlay for collection media
+         */
+        startMoveCollectionOverlay: function() {
+            this.sandbox.emit('sulu.media.collection-select.move-collection.open');
+        },
+
+        /**
+         * emit events to move collection
+         * @param collection
+         */
+        moveCollection: function(collection) {
+            this.sandbox.emit('sulu.media.collections.move', this.options.id, collection,
+                function() {
+                    var url = '/admin/api/collections/' + this.options.id + '?depth=1';
+
+                    this.sandbox.emit('husky.data-navigation.collections.set-url', url);
+                    this.sandbox.emit('sulu.labels.success.show', 'labels.success.collection-move-desc', 'labels.success');
+                }.bind(this)
+            );
+
+            this.sandbox.emit('sulu.media.collection-select.move-collection.restart');
+            this.sandbox.emit('sulu.media.collection-select.move-collection.close');
+        },
+
+        initialize: function() {
             this.locale = this.sandbox.sulu.user.locale;
             this.bindCustomEvents();
         },
 
-        bindCustomEvents: function () {
+        bindCustomEvents: function() {
             this.sandbox.on(SET_LOCALE.call(this), this.setLocale.bind(this));
             this.sandbox.on(GET_LOCALE.call(this), this.getLocale.bind(this));
+
+            // move collection overlay
+            this.sandbox.on('sulu.media.collection-select.move-collection.selected', this.moveCollection.bind(this));
         },
 
-        setLocale: function (locale) {
+        setLocale: function(locale) {
             this.locale = locale;
         },
 
-        getLocale: function (callback) {
+        getLocale: function(callback) {
             callback(this.locale);
         }
     };