+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/collections/components/root/main.js
@@ -1,0 +1,34 @@
+/*
+ * This file is part of the Sulu CMF.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+define(function() {
+
+    'use strict';
+
+    return {
+
+        view: true,
+
+        layout: {
+            navigation: {
+                collapsed: true
+            },
+            content: {
+                width: 'max'
+            }
+        },
+
+        /**
+         * Initializes the collections list
+         */
+        initialize: function() {
+            this.html('');
+        }
+    };
+});